Change colors to match current theme and move them to color package
