+++ core/ui/compose/theme/src/main/java/app/k9mail/core/ui/compose/theme/K9Theme.kt
@@ -2,6 +2,9 @@ package app.k9mail.core.ui.compose.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
+import app.k9mail.core.ui.compose.theme.color.MaterialColor
+import app.k9mail.core.ui.compose.theme.color.darkColors
+import app.k9mail.core.ui.compose.theme.color.lightColors
 
 private val k9LightColorPalette = lightColors(
     primary = MaterialColor.gray_800,
@@ -12,7 +15,7 @@ private val k9LightColorPalette = lightColors(
 
 private val k9DarkColorPalette = darkColors(
     primary = MaterialColor.gray_100,
-    primaryVariant = MaterialColor.gray_50,
+    primaryVariant = MaterialColor.gray_400,
     secondary = MaterialColor.pink_300,
     secondaryVariant = MaterialColor.pink_500,
 )
@@ -22,7 +25,7 @@ fun K9Theme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     content: @Composable () -> Unit,
 ) {
-    val images = Images(logo = R.drawable.core_ui_theme_logo_orange)
+    val images = Images(logo = R.drawable.core_ui_theme_k9_logo)
 
     MainTheme(
         lightColorPalette = k9LightColorPalette,
