+++ core/ui/compose/theme/src/main/res/drawable/core_ui_theme_k9_logo.xml
@@ -64,7 +64,7 @@
             android:pathData="M171.99,116V48C171.99,41.37 166.62,36 159.99,36L31.99,36C25.37,36 19.99,41.37 19.99,48V116C19.99,122.62 25.37,128 31.99,128L159.99,128C166.62,128 171.99,122.62 171.99,116Z"
             android:strokeAlpha="0.2" />
         <path
-            android:fillColor="#FFB74D"
+            android:fillColor="#FF2C55"
             android:pathData="M172,116V48C172,41.37 166.63,36 160,36L32,36C25.37,36 20,41.37 20,48V116C20,122.63 25.37,128 32,128H160C166.63,128 172,122.63 172,116Z" />
         <path
             android:fillColor="#00000000"
@@ -73,10 +73,10 @@
             android:strokeColor="#FBE9E7"
             android:strokeLineCap="round" />
         <path
-            android:fillColor="#FFB74D"
+            android:fillColor="#FF2C55"
             android:pathData="M32,36C25.35,36 20,41.35 20,48V49C20,42.35 25.35,37 32,37H160C166.65,37 172,42.35 172,49V48C172,41.35 166.65,36 160,36H32Z" />
         <path
-            android:fillColor="#F57C00"
+            android:fillColor="#C2185B"
             android:pathData="M20,115V116C20,122.65 25.35,128 32,128H160C166.65,128 172,122.65 172,116V115C172,121.65 166.65,127 160,127H32C25.35,127 20,121.65 20,115Z" />
         <path
             android:fillAlpha="0.2"
